+++ pkg/photo/photo_test.go
@@ -1,6 +1,7 @@
 package photo_test
 
 import (
+	"image/color"
 	"os"
 	"testing"
 
@@ -31,4 +32,51 @@ func TestPhotoInfo_Close(t *testing.T) {
 	require.NoError(t, err)
 	assert.NoError(t, pi.Close())
 	assert.Nil(t, pi.File)
+}
+
+func TestPhotoMeta_SetColor(t *testing.T) {
+	t.Parallel()
+
+	pm := photo.Meta{}
+	pm.SetColor(color.RGBA{0xaa, 0xbb, 0xcc, 0xff})
+	assert.Equal(t, "#aabbcc", pm.Color)
+}
+
+func TestPhotoMeta_GetColor(t *testing.T) {
+	t.Parallel()
+
+	pm := photo.Meta{Color: "#123456"}
+	c, err := pm.GetColor()
+	require.NoError(t, err)
+	assert.Equal(t, color.RGBA{0x12, 0x34, 0x56, 0xff}, c)
+
+	pm.Color = "123456"
+	c, err = pm.GetColor()
+	require.Error(t, err)
+	assert.Nil(t, c)
+
+	pm.Color = ""
+	c, err = pm.GetColor()
+	require.NoError(t, err)
+	assert.Nil(t, c)
+
+	pm.Color = "#12345"
+	c, err = pm.GetColor()
+	require.Error(t, err)
+	assert.Nil(t, c)
+
+	pm.Color = "#1x3456"
+	c, err = pm.GetColor()
+	require.Error(t, err)
+	assert.Nil(t, c)
+
+	pm.Color = "#12x456"
+	c, err = pm.GetColor()
+	require.Error(t, err)
+	assert.Nil(t, c)
+
+	pm.Color = "#12345x"
+	c, err = pm.GetColor()
+	require.Error(t, err)
+	assert.Nil(t, c)
 }