+++ pkg/text/esv/verse.go
@@ -1,16 +1,63 @@
 package esv
 
 import (
+	"context"
 	"html/template"
+	"net/url"
 
 	"github.com/zostay/go-esv-api/pkg/esv"
 
 	"github.com/zostay/today/pkg/ref"
+	"github.com/zostay/today/pkg/text"
 )
 
-// Verse returns the text of the given verse reference using the ESV API.
-func (r *Resolver) Verse(ref *ref.Resolved) (string, error) {
-	tr, err := r.Client.PassageText(ref.Ref(),
+// VersionInformation returns the metadata for the ESV from esv.org.
+func (r *Resolver) VersionInformation(context.Context) (*text.Version, error) {
+	return &text.Version{
+		Name: "ESV",
+		Link: "https://www.esv.org/",
+	}, nil
+}
+
+// Verse fetches a verse an associated metadata for the given reference.
+func (r *Resolver) Verse(ctx context.Context, ref *ref.Resolved) (*text.Verse, error) {
+	txt, err := r.VerseText(ctx, ref)
+	if err != nil {
+		return nil, err
+	}
+
+	html, err := r.VerseHTML(ctx, ref)
+	if err != nil {
+		return nil, err
+	}
+
+	path, err := url.JoinPath(
+		"https://www.esv.org/",
+		url.PathEscape(ref.Ref()),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	vi, err := r.VersionInformation(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	return &text.Verse{
+		Reference: ref.Ref(),
+		Content: text.Content{
+			Text: txt,
+			HTML: html,
+		},
+		Link:    path,
+		Version: *vi,
+	}, nil
+}
+
+// VerseText returns the text of the given verse reference using the ESV API.
+func (r *Resolver) VerseText(ctx context.Context, ref *ref.Resolved) (string, error) {
+	tr, err := r.Client.PassageTextContext(ctx, ref.Ref(),
 		esv.WithIncludeVerseNumbers(false),
 		esv.WithIncludeHeadings(false),
 		esv.WithIncludeFootnotes(false),
@@ -23,8 +70,8 @@ func (r *Resolver) Verse(ref *ref.Resolved) (string, error) {
 }
 
 // VerseHTML returns the HTML of the given verse reference using the ESV API.
-func (r *Resolver) VerseHTML(ref *ref.Resolved) (template.HTML, error) {
-	tr, err := r.Client.PassageHtml(ref.Ref(),
+func (r *Resolver) VerseHTML(ctx context.Context, ref *ref.Resolved) (template.HTML, error) {
+	tr, err := r.Client.PassageHtmlContext(ctx, ref.Ref(),
 		esv.WithIncludeVerseNumbers(false),
 		esv.WithIncludeHeadings(false),
 		esv.WithIncludeFootnotes(false),
@@ -40,3 +87,5 @@ func (r *Resolver) VerseHTML(ref *ref.Resolved) (template.HTML, error) {
 
 	return template.HTML(tr.Passages[0]), nil //nolint:gosec // we trust the ESV API
 }
+
+var _ text.Resolver = (*Resolver)(nil)